+++ CHANGELOG.md
@@ -5,9 +5,8 @@
 ## Fixed
 
 - Goole Big Query:
-    - fix Validation error when all GoogleCredentials are not set (only the `project_id` is required).
-    - Explicit error information when no data to return.
-    - fallback on normal GoogleCredentials connection mode when jwt-token is not valid aymore.
+    - Explicit errors information when no data to return.
+    - Fallback on normal GoogleCredentials connection when JWTCredentials fails (jwt-token is not valid aymore).
 
 ### [4.9.0] 2023-09-20