+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ## Fixed
 
 - Goole Big Query:
-    - Explicit errors information when no data to return.
-    - Fallback on normal GoogleCredentials connection when JWTCredentials fails (jwt-token is not valid aymore).
+    - Better UX (Switch between GoogleCreds auth or GoogleJWT  auth).
+    - Explicit errors information when no data is returned.
+    - Fallback on GoogleCredentials auth when JWTCredentials fails (or when jwt-token is not valid aymore).
 
 ### [4.9.0] 2023-09-20