+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+### [4.9.1] 2023-09-22
+
 ## Fixed
 
 - Goole Big Query: